fix: BinaryCOSNWriter 에서 숫자를 원본 타입을 따라가지 않고 단지 크기에 따라서 쓰기 공간을 사용하도록 변경

+++ src/test/java/com/hancomins/cson/DefaultUse.java
@@ -17,13 +17,13 @@ public class DefaultUse {
     @Test
     public void toBinaryArrayTest() {
         CSONObject csonObject = new CSONObject();
-        csonObject.put("number", 1000).put("bigNumber", new BigDecimal(10000000000L));
+        csonObject.put("number", (short)1000).put("bigNumber", new BigDecimal(10000000000L));
 
         byte[] bytes = csonObject.toBytes();
 
         CSONObject parsedCsonObject = new CSONObject(bytes);
         assertEquals(csonObject.get("number"), parsedCsonObject.get("number"));
-        assertEquals(1000, parsedCsonObject.get("number"));
+        assertEquals((short)1000, parsedCsonObject.get("number"));
 
         assertEquals(10000000000L, parsedCsonObject.getLong("bigNumber"));
